Update Portuguese.properties
Hello! I have just found out this game, and i'd like to contribute to its project. As portuguese is my first language, i'd like to start contributing with the translation of the game. I made some changes here and there, but i'm looking forward to keep translating more of it soon.